Add files via upload
TASK-1A
Artificial neural networks (ANNs) and deep neural networks use backpropagation as a learning algorithm to compute a gradient descent, which is an optimization algorithm that guides the user to the maximum or minimum of a function.

In a machine learning context, the gradient descent helps the system minimize the gap between desired outputs and achieved system outputs. The algorithm tunes the system by adjusting the weight values for various inputs to narrow the difference between outputs. This is also known as the error between the two.

More specifically, a gradient descent algorithm uses a gradual process to provide information on how a network's parameters need to be adjusted to reduce the disparity between the desired and achieved outputs. An evaluation metric called a cost function guides this process. The cost function is a mathematical function that measures this error. The algorithm's goal is to determine how the parameters must be adjusted to reduce the cost function and improve overall accuracy.

Backpropagation, or backward propagation of errors, is an algorithm that is designed to test for errors working back from output nodes to input nodes. It's an important mathematical tool for improving the accuracy of predictions in data mining and machine learning. Essentially, backpropagation is an algorithm used to quickly calculate derivatives in a neural network, which are the changes in output because of tuning and adjustments.

There are two leading types of backpropagation networks:

Static backpropagation. Static backpropagation is a network developed to map static inputs for static outputs. Static networks can solve static classification problems, such as optical character recognition (OCR).
Recurrent backpropagation. The recurrent backpropagation network is used for fixed-point learning. This means that during neural network training, the weights are numerical values that determine how much nodes -- also referred to as neurons -- influence output values. They're adjusted so that the network can achieve stability by reaching a fixed value.