Add files via upload
TASK-2
From telling rickshaw-wala where to go, to tell him where to come we have grown up. Yes, we are talking about online cab and bike facility providers like OLA ,Rapido and Uber. If you had used this app some times then you must have paid some day less and someday more for the same journey. But have you ever thought what is the reason behind it? It is because of the high demand at some hours. this is not the only factor but this is one of them.

Rapido Bike Ride Request Forecast using ML

In this task we will try to predict ride-request for a particular hour using machine learning. One can refer to the below explanation for the column names in the dataset and their values as well.

season:
spring
summer
fall
winter
weather

1)Clear, Few clouds, Partly cloudy, Partly cloudy
2)Mist + Cloudy, Mist + Broken clouds, Mist + Few clouds, Mist
3)Light Snow, Light Rain + Thunderstorm + Scattered clouds, Light Rain + Scattered clouds
4)Heavy Rain + Ice Pallets + Thunderstorm + Mist, Snow + Fog

casual – number of non-registered user rentals initiated
registered – number of registered user rentals initiated
count – number of ride request raised on the app for that particular hour.